Move enum util to homeassistant.util
