+++ homeassistant/helpers/device_registry.py
@@ -11,9 +11,9 @@ import attr
 from homeassistant.const import EVENT_HOMEASSISTANT_STARTED
 from homeassistant.core import Event, HomeAssistant, callback
 from homeassistant.exceptions import RequiredParameterMissing
-from homeassistant.helpers.enum import StrEnum
 from homeassistant.helpers.frame import report
 from homeassistant.loader import bind_hass
+from homeassistant.util.enum import StrEnum
 import homeassistant.util.uuid as uuid_util
 
 from .debounce import Debouncer
